requirements.txt
numpy>=1.24
pandas>=2.0
pyyaml>=6.0
requests>=2.31
scikit-learn>=1.3
lightgbm>=4.0
joblib>=1.3
matplotlib>=3.7
vaderSentiment>=3.3.2
transformers>=4.41
torch>=2.2
